Fix for final step of site registration failing in SitesController (#1131)
* Rename file to match contents

* Test attempt to send notification emails when no remote users for site

* Fix in situation where no remote users
